[BREAKING] Make PluginManager an interface and let dispatchCommand() throw exceptions.
The current BungeeCord implementation of command dispatching is utterly inappropriate for plugins based on successful command execution (i.e. Buycraft). This fixes the issue by allowing an exception to be thrown.

Additionally, PluginManager has become an interface. This allows more flexibility when major changes (like this change) are made.
