+++ proxy/src/main/java/Test.java
@@ -2,7 +2,10 @@
 import net.md_5.bungee.BungeeCord;
 import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.ProxyServer;
+import net.md_5.bungee.api.plugin.CommandExecutionException;
 import net.md_5.bungee.command.ConsoleCommandSender;
+
+import java.util.logging.Level;
 
 /*
  * To change this template, choose Tools | Templates
@@ -27,9 +30,15 @@ public class Test
             String line = bungee.getConsoleReader().readLine( ">" );
             if ( line != null )
             {
-                if ( !bungee.getPluginManager().dispatchCommand( ConsoleCommandSender.getInstance(), line ) )
+                try
                 {
-                    bungee.getConsole().sendMessage( ChatColor.RED + "Command not found" );
+                    if ( !bungee.getPluginManager().dispatchCommand( ConsoleCommandSender.getInstance(), line ) )
+                    {
+                        bungee.getConsole().sendMessage(ChatColor.RED + "Command not found");
+                    }
+                } catch ( CommandExecutionException e )
+                {
+                    bungee.getLogger().log( Level.WARNING, "Error in dispatching command", e );
                 }
             }
         }
